+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Output improvements with log messages and progress bars
 * Small bug fixes when generating properties
 * Refactor to parallelize the initial data read more
+* Support writing to GeoJSONSeq format files, which tippecanoe can read much
+  faster
 
 ## v0.11.0 (2023-10-18)